feat(e2e tests): Add python-based end to end tests starter
Includes:
Tests based on the behaviour driven testing model with pytest-bdd as the package providing the helpers.
Each test represents a scenario.
Scenarios are found in a features file
Each contract would eventually have a features file in which a number of scenarios are defined
Each scenario has a number of steps
A fully tested scenario is represented as a python file which implements the steps and verifies the expected requirements.