Fix `astro_lp_address` usage in contract
Throughout the contract there is inconsistencies as to what the meaning
of `state.astro_lp_address` and `InstantiateMsg.astro_lp_address`
actually refer to. In some places, we refer to it as the *token*
returned by the LP for providing liquidity, while in others (namely
`deposit_passive_strategy`), we refer to it as the bLuna-Luna pool
address.

This commit rectifies the inconsistencies and declares the
`astro_lp_address` to be the address of the bLuna-Luna LP pool. To get
the liquid token returned for providing assets, there is a helper
function `get_lp_token_address` which can be used.